fix(gridster): Orphan preview holder when dragging is interrupted
After beginning a drag, if the mouse button is released outside of the
screen, and if a subsequent drag operation is begun on another widget
(while the previous widget still follows the mouse around), Draggable
will trigger a new drag start event, causing Gridster to leave the
current widget + its preview holder aside and start dragging the new
widget. This will lead to an orphan preview holder to be left on the
screen.

A quick solution is to ignore drag start without receiving a drag stop.